requirements.txt
requests
beautifulsoup4
pillow
pandas
selenium
webdriver-manager
